+++ src/command/commands/personality.ts
@@ -4,6 +4,7 @@ const command: Command = {
   name: 'personality',
   platforms: [Platform.Twitch, Platform.Discord],
   run: async (ctx, { commandState }) => {
+    console.log('here');
     if (ctx.tags.owner && ctx.body) {
       commandState.blunderBotPersonality = ctx.body.trim();
     }