+++ src/twitch/twitch.service.ts
@@ -4,10 +4,10 @@ import { TwitchGateway } from './twitch.gateway';
 import { CommandService } from '../command/command.service';
 import { writeLog } from '../utils/logs';
 import { Platform } from '../enums';
-import { TwitchCustomRewardsService } from './twitch.custom-rewards';
 import { chessSquares } from '../utils/constants';
 import { getRandomElement } from '../utils/utils';
-const tmi = require('tmi.js');
+import WebSocket = require('ws');
+import tmi = require('tmi.js');
 
 let shoutoutUsers = CONFIG.autoShoutouts;
 const newChatters = [];
@@ -42,8 +42,6 @@ export class TwitchService {
   constructor(
     @Inject(forwardRef(() => CommandService))
     private readonly commandService: CommandService,
-    @Inject(forwardRef(() => TwitchCustomRewardsService))
-    private readonly twitchCustomRewardsService: TwitchCustomRewardsService,
     @Inject(forwardRef(() => TwitchGateway))
     private readonly twitchGateway: TwitchGateway
   ) {
@@ -61,6 +59,28 @@ export class TwitchService {
     this.client.on('resub', this.onResubHandler.bind(this));
     this.client.on('subgift', this.onSubGiftHandler.bind(this));
     this.client.on('submysterygift', this.onSubMysteryGiftHandler.bind(this));
+
+    // TODO: Eventually move everything to PubSub
+    const pubSubConnection = new WebSocket('wss://pubsub-edge.twitch.tv');
+    pubSubConnection.onopen = () => {
+      this.logger.log('PubSub connection opened');
+      pubSubConnection.send(
+        JSON.stringify({
+          type: 'LISTEN',
+          data: {
+            topics: [
+              `channel-points-channel-v1.${CONFIG.twitch.ownerId}`,
+              `channel-subscribe-events-v1.${CONFIG.twitch.ownerId}`,
+              `channel-bits-badge-unlocks.${CONFIG.twitch.ownerId}`,
+              `channel-bits-events-v2.${CONFIG.twitch.ownerId}`
+            ],
+            auth_token: CONFIG.twitch.apiOwnerOauthToken
+          }
+        })
+      );
+    };
+
+    pubSubConnection.onmessage = this.pubSubMessageHandler.bind(this);
   }
 
   botSpeak(message: string) {
@@ -81,6 +101,33 @@ export class TwitchService {
 
   updateBoughtSquares(data) {
     boughtSquares = data;
+  }
+
+  async pubSubMessageHandler(data) {
+    const event = JSON.parse(data.data);
+
+    if (event.type === 'MESSAGE') {
+      const message = JSON.parse(event.data.message);
+      console.log(message);
+
+      if (message.type === 'reward-redeemed') {
+        const username = message.data.redemption.user.display_name;
+        const userInput = message.data.redemption.user_input;
+        const reward = message.data.redemption.reward;
+
+        this.logger.log(`User ${username} redeemed ${reward.title}`);
+
+        if (CONFIG.twitch.customRewardCommands[reward.title]) {
+          for (let command of CONFIG.twitch.customRewardCommands[
+            reward.title
+          ]) {
+            command = command.replace(/{username}/gi, `${username}`);
+            command = command.replace(/{message}/gi, `${userInput}`);
+            void this.ownerRunCommand(`${command}`);
+          }
+        }
+      }
+    }
   }
 
   async onMessageHandler(channel, tags, message) {
@@ -143,9 +190,7 @@ export class TwitchService {
       return;
     }
 
-    if (tags['custom-reward-id']) {
-      await this.twitchCustomRewardsService.handleCustomRewards(context);
-    } else {
+    if (!tags['custom-reward-id']) {
       await this.commandService.run(context);
     }
   }