+++ RELEASE_NOTES.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+### 1.3.0 (2024-01-04)
+
+#### Breaking Changes
+
+- PUBSUB is here!
+- Removed all the custom rewards functionality from the previous iteration. All custom rewards are now under `twitch.customRewardCommands`. Getting the custom reward id is no longer necessary.
+
+#### Features
+
+- Added an `!opp-rating` owner command to support the new custom reward changes.
+- Edited the `!queue` command with "add" support for owners to support the new custom reward changes and moved the queue to command state.
+
 ### 1.2.0 (2024-01-02)
 
 #### Breaking Changes